requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
dj-database-url
whitenoise==6.9.0
django-cors-headers==4.7.0
gunicorn
